feat(required-plugins): initialized the TGMPA library
Added a check for the TGMPA class, loaded it, and add a new action for
initializing TGMPA. Moved the Advanced Custom Fields dependency to
TGMPA in the child theme in favor of bundling with Hoverboard, which
doesn't actually require ACF to operate.

Added a check for WP's debug mode and activated ACF_LITE if it's off.