feat(modularization): set up main.less to include stylesheets
In an effort to make styles easier to dig through later, let's
break them into smaller files with higher specificity.

For example, put all header/nav styles in `header.less`, all
typography rules for the site in `typography.less`, and then
put layout rules in files that match the templating: `single.less`
and so on.